+++ src/components/pages/ClientDashboard/index.js
@@ -1,1 +1,1 @@
-export { default as ClientDash } from './RenderClientDashboard';
+export { default as ClientDash } from './ClientDashboardContainer.js';
